refactor(routes): extract tool helpers and platform-attempts runner
Move shared chat.py infrastructure into format-agnostic modules so the
upcoming /v1/messages and /v1/responses parity work can reuse them
without duplicating logic.

- _tools.py: tool extraction + gateway-field stripping + per-tool
  config resolution (8 helpers, no caller changes).
- _platform.py: ResolvedRoute / ResolvedAttempt types and a generic
  run_platform_attempts runner. Response type via TypeVar T;
  per-attempt dispatch, usage extraction, error classification, success
  hook, and background outcome reporting are injected by the caller.
  Lock-in semantics: run_attempt receives a per-attempt
  on_first_response callback (matches the contract introduced in #73)
  that tool-loop callers thread into mcp_tool_loop. Once that callback
  has fired, any subsequent error from the attempt terminates the
  request instead of falling through, since tool-loop transcripts can't
  be transparently replayed on a different provider.
- chat.py: imports back from the new modules; the non-stream platform
  branch now builds a _run_chat_attempt closure (with the same MCP /
  sandbox / web_search / plain dispatch as before, now threading
  on_first_response through to mcp_tool_loop) and calls
  run_platform_attempts instead of running the attempts loop inline.

No behavior change relative to main. Includes the multi-PR plan
documenting the staged migration for Anthropic messages, OpenAI
Responses, and platform-mode routing.
