+++ src/gateway/api/routes/chat.py
@@ -5,7 +5,7 @@ import uuid
 from collections.abc import AsyncIterator, Callable
 from contextlib import AsyncExitStack
 from datetime import UTC, datetime
-from typing import Annotated, Any, NamedTuple
+from typing import Annotated, Any
 
 import httpx
 from any_llm import AnyLLM, LLMProvider, acompletion
@@ -21,6 +21,14 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from gateway.api.deps import get_config, get_db_if_needed, get_log_writer, verify_api_key_or_master_key
 from gateway.api.routes._helpers import resolve_user_id
+from gateway.api.routes._platform import ResolvedAttempt, ResolvedRoute, run_platform_attempts
+from gateway.api.routes._tools import (
+    _build_web_search_backend,
+    _extract_code_execution_tool,
+    _extract_web_search_tool,
+    _resolve_sandbox_purpose_hint,
+    _strip_gateway_fields,
+)
 from gateway.core.config import GatewayConfig
 from gateway.log_config import logger
 from gateway.metrics import record_cost, record_tokens
@@ -41,7 +49,7 @@ from gateway.services.mcp_loop import (
 from gateway.services.pricing_service import find_model_pricing
 from gateway.services.provider_kwargs import get_provider_kwargs as get_provider_kwargs  # noqa: F401
 from gateway.services.sandbox_backend import SandboxBackend, SandboxNotReachableError
-from gateway.services.web_search_backend import WebSearchBackend, WebSearchNotReachableError
+from gateway.services.web_search_backend import WebSearchNotReachableError
 from gateway.streaming import (
     OPENAI_STREAM_FORMAT,
     StreamingAttemptFailure,
@@ -70,206 +78,6 @@ def rate_limit_headers(info: RateLimitInfo) -> dict[str, str]:
         "X-RateLimit-Remaining": str(info.remaining),
         "X-RateLimit-Reset": str(int(info.reset)),
     }
-
-
-def _is_web_search_tool_type(type_value: Any) -> bool:
-    """Recognise the tool-array shapes that map to the web_search backend.
-
-    Accepts:
-      * ``"web_search"`` — gateway-native short form (matches OpenAI's
-        ``{"type": "web_search"}`` server-managed tool).
-      * ``"web_search_*"`` — Anthropic versioned types (e.g.
-        ``"web_search_20250305"``, future ``"web_search_20260209"``).
-
-    Matching Anthropic by prefix means new versions keep working without a
-    code change; the backend's search semantics are version-agnostic.
-    """
-    if not isinstance(type_value, str):
-        return False
-    return type_value == "web_search" or type_value.startswith("web_search_")
-
-
-def _is_code_execution_tool_type(type_value: Any) -> bool:
-    """Recognise the tool-array shapes Anthropic and OpenAI use for code execution.
-
-    Accepts:
-      * ``"code_execution"`` — gateway-native short form
-      * ``"code_interpreter"`` — OpenAI Responses/Assistants API
-      * ``"code_execution_*"`` — Anthropic versioned types
-        (e.g. ``"code_execution_20250825"``)
-
-    Matching Anthropic by prefix means new versions (``code_execution_20260101``,
-    etc.) keep working without a code change. Our sandbox is a generic Python
-    REPL, so we don't need to track per-version semantics.
-    """
-    if not isinstance(type_value, str):
-        return False
-    return (
-        type_value == "code_execution" or type_value == "code_interpreter" or type_value.startswith("code_execution_")
-    )
-
-
-# Gateway-internal fields the provider SDKs (any-llm, anthropic, openai, …)
-# don't accept as ``acompletion`` kwargs. Strip these from the model_dump
-# before forwarding to upstream — Anthropic in particular rejects unknown
-# kwargs with a hard error.
-_GATEWAY_INTERNAL_FIELDS = (
-    "mcp_servers",
-    "mcp_server_ids",
-    "tools_header",
-    "max_tool_iterations",
-    "user",
-)
-
-
-def _strip_gateway_fields(
-    fields: dict[str, Any],
-    *,
-    tools_extracted: bool = False,
-    remaining_user_tools: list[dict[str, Any]] | None = None,
-) -> dict[str, Any]:
-    """Strip gateway-internal fields from a ``request.model_dump(...)`` payload.
-
-    Mutates ``fields`` in place and returns it for chaining. When the caller
-    extracted any gateway-managed tool entry from ``tools`` (sandbox /
-    web_search / future), pass ``tools_extracted=True`` and the remaining
-    user-supplied tools; the original ``tools`` list is replaced (or popped
-    entirely if none remain).
-    """
-    for k in _GATEWAY_INTERNAL_FIELDS:
-        fields.pop(k, None)
-    if tools_extracted:
-        if remaining_user_tools:
-            fields["tools"] = remaining_user_tools
-        else:
-            fields.pop("tools", None)
-    return fields
-
-
-def _resolve_sandbox_purpose_hint(sandbox_tool_entry: dict[str, Any] | None) -> str | None:
-    """Resolve the per-tool ``purpose_hint`` for the sandbox.
-
-    Priority: tool entry's ``purpose_hint`` → ``GATEWAY_SANDBOX_PURPOSE_HINT``
-    env → ``None`` (SandboxBackend falls back to its built-in default).
-    """
-    return (
-        (sandbox_tool_entry.get("purpose_hint") if sandbox_tool_entry else None)
-        or os.environ.get("GATEWAY_SANDBOX_PURPOSE_HINT")
-        or None
-    )
-
-
-def _extract_first_matching_tool(
-    tools: list[dict[str, Any]] | None,
-    predicate: Callable[[Any], bool],
-) -> tuple[dict[str, Any] | None, list[dict[str, Any]] | None]:
-    """Pull the first tool entry whose ``type`` matches ``predicate``.
-
-    Returns ``(entry_or_None, remaining_tools_or_None)``. The extracted entry
-    is thin (no function schema); the gateway-managed backend's
-    ``openai_tools`` provides the full definition during tool-use-loop
-    injection. Remaining user-supplied tools pass through unchanged.
-    """
-    if not tools:
-        return None, tools
-    entry: dict[str, Any] | None = None
-    remaining: list[dict[str, Any]] = []
-    for t in tools:
-        if entry is None and isinstance(t, dict) and predicate(t.get("type")):
-            entry = t
-        else:
-            remaining.append(t)
-    return entry, (remaining or None)
-
-
-def _extract_code_execution_tool(
-    tools: list[dict[str, Any]] | None,
-) -> tuple[dict[str, Any] | None, list[dict[str, Any]] | None]:
-    """Pull the first code-execution-style entry out of ``tools``.
-
-    Detects the gateway-native ``{"type": "code_execution"}`` shape plus the
-    provider-native equivalents from OpenAI (``code_interpreter``) and Anthropic
-    (``code_execution_20250825`` and future versions). All three map to the same
-    sandbox backend so swapping ``base_url`` to the gateway keeps existing
-    SDK code working unchanged.
-    """
-    return _extract_first_matching_tool(tools, _is_code_execution_tool_type)
-
-
-def _extract_web_search_tool(
-    tools: list[dict[str, Any]] | None,
-) -> tuple[dict[str, Any] | None, list[dict[str, Any]] | None]:
-    """Pull the first web-search-style entry out of ``tools``.
-
-    Accepts both the gateway-native ``{"type": "web_search"}`` shape (matching
-    OpenAI's server-managed tool) and Anthropic's dated variants
-    (``web_search_20250305`` etc.). All map to the same backend.
-    """
-    return _extract_first_matching_tool(tools, _is_web_search_tool_type)
-
-
-def _resolve_web_search_purpose_hint(tool_entry: dict[str, Any] | None) -> str | None:
-    """Per-tool entry → ``GATEWAY_WEB_SEARCH_PURPOSE_HINT`` → ``None`` (backend default)."""
-    return (
-        (tool_entry.get("purpose_hint") if tool_entry else None)
-        or os.environ.get("GATEWAY_WEB_SEARCH_PURPOSE_HINT")
-        or None
-    )
-
-
-def _build_web_search_backend(*, base_url: str, tool_entry: dict[str, Any]) -> WebSearchBackend:
-    """Construct a WebSearchBackend honouring env-level + per-tool config.
-
-    Per-tool entry fields (``max_results``, ``allowed_domains``,
-    ``blocked_domains``, ``purpose_hint``) override env-level defaults.
-    Operator-level env knobs:
-
-      * ``GATEWAY_WEB_SEARCH_ENGINES`` — comma-separated SearXNG engine list
-      * ``GATEWAY_WEB_SEARCH_MAX_RESULTS`` — default cap on returned hits
-      * ``GATEWAY_WEB_SEARCH_EXTRACT`` — "0"/"false" to disable in-process
-        content extraction (snippet-only mode).
-      * ``GATEWAY_WEB_SEARCH_PURPOSE_HINT`` — per-deployment hint override.
-    """
-    kwargs: dict[str, Any] = {"base_url": base_url}
-
-    engines_str = os.environ.get("GATEWAY_WEB_SEARCH_ENGINES")
-    if engines_str:
-        engines = tuple(e.strip() for e in engines_str.split(",") if e.strip())
-        if engines:
-            kwargs["engines"] = engines
-
-    max_env = os.environ.get("GATEWAY_WEB_SEARCH_MAX_RESULTS")
-    if max_env:
-        try:
-            parsed_max = int(max_env)
-        except ValueError:
-            logger.warning("GATEWAY_WEB_SEARCH_MAX_RESULTS=%r is not an int; ignoring", max_env)
-        else:
-            if parsed_max >= 1:
-                kwargs["max_results"] = parsed_max
-            else:
-                logger.warning("GATEWAY_WEB_SEARCH_MAX_RESULTS=%r is not >= 1; ignoring", max_env)
-    req_max = tool_entry.get("max_results")
-    if isinstance(req_max, int) and req_max > 0:
-        kwargs["max_results"] = req_max
-
-    extract_env = os.environ.get("GATEWAY_WEB_SEARCH_EXTRACT")
-    if extract_env is not None:
-        kwargs["extract_content"] = extract_env.lower() not in {"0", "false", "no", "off"}
-
-    allowed = tool_entry.get("allowed_domains")
-    if isinstance(allowed, list) and allowed:
-        kwargs["allowed_domains"] = tuple(str(d) for d in allowed)
-    blocked = tool_entry.get("blocked_domains")
-    if isinstance(blocked, list) and blocked:
-        kwargs["blocked_domains"] = tuple(str(d) for d in blocked)
-
-    purpose_hint = _resolve_web_search_purpose_hint(tool_entry)
-    if purpose_hint:
-        kwargs["purpose_hint"] = purpose_hint
-
-    return WebSearchBackend(**kwargs)
-
 
 class ChatCompletionRequest(BaseModel):
     """OpenAI-compatible chat completion request."""
@@ -427,26 +235,6 @@ async def _post_platform(
 ) -> httpx.Response:
     async with httpx.AsyncClient(timeout=timeout_seconds) as client:
         return await client.post(url, headers=headers, json=body)
-
-
-class ResolvedAttempt(BaseModel):
-    """A single resolution attempt returned by the platform."""
-
-    attempt_id: str
-    position: int
-    provider: str
-    model: str
-    api_base: str | None = None
-    api_key: str
-    managed: bool
-
-
-class ResolvedRoute(BaseModel):
-    """The full resolution plan returned by the platform."""
-
-    request_id: str
-    fallback_enabled: bool
-    attempts: list[ResolvedAttempt]
 
 
 async def _resolve_platform_credentials(
@@ -1157,204 +945,109 @@ async def chat_completions(
                 status_code=status.HTTP_502_BAD_GATEWAY,
                 detail="Authorization service returned no resolvable provider",
             )
-    else:
-        provider, model = AnyLLM.split_model_provider(request.model)
-        provider_kwargs = get_provider_kwargs(config, provider)
-        attempts_to_try = []  # standalone path doesn't use the attempts list
 
-    class _AttemptFailure(NamedTuple):
-        position: int
-        provider: str
-        model: str
-        error_class: str
-
-    failures: list[_AttemptFailure] = []
-    last_exc: BaseException | None = None
-
-    if platform_mode:
         base_request_fields = _strip_gateway_fields(
             request.model_dump(exclude_unset=True),
             tools_extracted=sandbox_tool_entry is not None or web_search_tool_entry is not None,
             remaining_user_tools=remaining_user_tools,
         )
-        for attempt in attempts_to_try:
-            attempt_provider = LLMProvider(attempt.provider)
-            attempt_model = attempt.model
-            attempt_kwargs: dict[str, Any] = {"api_key": attempt.api_key}
-            if attempt.api_base:
-                attempt_kwargs["api_base"] = attempt.api_base
 
-            completion_kwargs = {
-                **attempt_kwargs,
-                **base_request_fields,
-                "model": f"{attempt_provider.value}:{attempt_model}",
-            }
+        async def _run_chat_attempt(
+            completion_kwargs: dict[str, Any],
+            on_first_response: Callable[[], None],
+        ) -> ChatCompletion:
+            if mcp_server_configs:
+                async with MCPClientPool(mcp_server_configs) as pool:
+                    mcp_kwargs = {
+                        **completion_kwargs,
+                        "messages": inject_purpose_hints(
+                            completion_kwargs["messages"],
+                            pool.purpose_hints(),
+                            header=request.tools_header,
+                        ),
+                    }
+                    return await mcp_tool_loop(
+                        completion_kwargs=mcp_kwargs,
+                        pool=pool,
+                        max_iterations=max_tool_iterations,
+                        on_first_response=on_first_response,
+                    )
+            if use_sandbox:
+                assert sandbox_url is not None
+                sandbox_hint = _resolve_sandbox_purpose_hint(sandbox_tool_entry)
+                async with SandboxBackend(sandbox_url=sandbox_url, purpose_hint=sandbox_hint) as backend:
+                    sandbox_kwargs = {
+                        **completion_kwargs,
+                        "messages": inject_purpose_hints(
+                            completion_kwargs["messages"],
+                            backend.purpose_hints(),
+                            header=request.tools_header,
+                        ),
+                    }
+                    return await mcp_tool_loop(
+                        completion_kwargs=sandbox_kwargs,
+                        pool=backend,  # type: ignore[arg-type]
+                        max_iterations=max_tool_iterations,
+                        on_first_response=on_first_response,
+                    )
+            if use_web_search:
+                assert web_search_url is not None
+                assert web_search_tool_entry is not None
+                async with _build_web_search_backend(
+                    base_url=web_search_url,
+                    tool_entry=web_search_tool_entry,
+                ) as web_backend:
+                    web_kwargs = {
+                        **completion_kwargs,
+                        "messages": inject_purpose_hints(
+                            completion_kwargs["messages"],
+                            web_backend.purpose_hints(),
+                            header=request.tools_header,
+                        ),
+                    }
+                    return await mcp_tool_loop(
+                        completion_kwargs=web_kwargs,
+                        pool=web_backend,  # type: ignore[arg-type]
+                        max_iterations=max_tool_iterations,
+                        on_first_response=on_first_response,
+                    )
+            return await acompletion(**completion_kwargs)  # type: ignore[return-value]
 
-            # Per-attempt lock-in flag. Flipped the moment the upstream
-            # returns its first assistant message. After that, any failure
-            # terminates the request — fallback would replay a provider-
-            # specific transcript on a different provider, which has
-            # undefined semantics.
-            locked_in = False
-
-            def _mark_locked_in(_pos: int = attempt.position) -> None:
-                nonlocal locked_in
-                locked_in = True
-                logger.info(
-                    "Tool-loop lock-in request_id=%s position=%d provider=%s model=%s",
-                    platform_route.request_id,
-                    _pos,
-                    attempt.provider,
-                    attempt.model,
-                )
-
-            try:
-                if mcp_server_configs:
-                    async with MCPClientPool(mcp_server_configs) as pool:
-                        mcp_kwargs = {
-                            **completion_kwargs,
-                            "messages": inject_purpose_hints(
-                                completion_kwargs["messages"],
-                                pool.purpose_hints(),
-                                header=request.tools_header,
-                            ),
-                        }
-                        completion: ChatCompletion = await mcp_tool_loop(
-                            completion_kwargs=mcp_kwargs,
-                            pool=pool,
-                            max_iterations=max_tool_iterations,
-                            on_first_response=_mark_locked_in,
-                        )
-                elif use_sandbox:
-                    assert sandbox_url is not None
-                    sandbox_hint = _resolve_sandbox_purpose_hint(sandbox_tool_entry)
-                    async with SandboxBackend(sandbox_url=sandbox_url, purpose_hint=sandbox_hint) as backend:
-                        sandbox_kwargs = {
-                            **completion_kwargs,
-                            "messages": inject_purpose_hints(
-                                completion_kwargs["messages"],
-                                backend.purpose_hints(),
-                                header=request.tools_header,
-                            ),
-                        }
-                        completion = await mcp_tool_loop(
-                            completion_kwargs=sandbox_kwargs,
-                            pool=backend,  # type: ignore[arg-type]
-                            max_iterations=max_tool_iterations,
-                            on_first_response=_mark_locked_in,
-                        )
-                elif use_web_search:
-                    assert web_search_url is not None
-                    assert web_search_tool_entry is not None
-                    async with _build_web_search_backend(
-                        base_url=web_search_url,
-                        tool_entry=web_search_tool_entry,
-                    ) as web_backend:
-                        web_kwargs = {
-                            **completion_kwargs,
-                            "messages": inject_purpose_hints(
-                                completion_kwargs["messages"],
-                                web_backend.purpose_hints(),
-                                header=request.tools_header,
-                            ),
-                        }
-                        completion = await mcp_tool_loop(
-                            completion_kwargs=web_kwargs,
-                            pool=web_backend,  # type: ignore[arg-type]
-                            max_iterations=max_tool_iterations,
-                            on_first_response=_mark_locked_in,
-                        )
-                else:
-                    completion = await acompletion(**completion_kwargs)  # type: ignore[assignment]
-            except HTTPException:
-                raise
-            except MaxToolIterationsExceeded as exc:
-                # The gateway's own iteration cap was hit, not an upstream
-                # failure. Surface a distinct 422 so callers can tell a
-                # runaway tool loop apart from a provider outage.
-                logger.warning(
-                    "Tool loop iteration cap hit request_id=%s position=%d cap=%d",
-                    platform_route.request_id,
-                    attempt.position,
-                    max_tool_iterations,
-                )
-                raise HTTPException(
-                    status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-                    detail=str(exc),
-                ) from exc
-            except BaseException as exc:
-                retryable, error_class = _classify_upstream_error(exc)
-                background_tasks.add_task(
-                    _report_platform_usage,
-                    config,
-                    attempt.attempt_id,
-                    "error",
-                    None,
-                    error_class,
-                )
-                logger.warning(
-                    "Provider call failed request_id=%s position=%d provider=%s model=%s "
-                    "error=%s retryable=%s locked_in=%s",
-                    platform_route.request_id,
-                    attempt.position,
-                    attempt.provider,
-                    attempt.model,
-                    error_class,
-                    retryable,
-                    locked_in,
-                )
-                last_exc = exc
-                # Locked-in: at least one tool-loop round produced an
-                # assistant message on this attempt. Subsequent failures
-                # cannot be transparently retried on another provider — the
-                # conversation now contains provider-specific tool_call ids
-                # and reasoning blocks. Surface immediately.
-                if locked_in:
-                    raise HTTPException(
-                        status_code=status.HTTP_502_BAD_GATEWAY,
-                        detail="LLM provider error",
-                    ) from exc
-                if not retryable:
-                    raise HTTPException(
-                        status_code=status.HTTP_502_BAD_GATEWAY,
-                        detail="LLM provider error",
-                    ) from exc
-                failures.append(_AttemptFailure(attempt.position, attempt.provider, attempt.model, error_class))
-                continue
-
-            # Success on this attempt.
+        def _report_attempt_outcome(
+            attempt: ResolvedAttempt,
+            outcome: str,
+            usage: Any,
+            error_class: str | None,
+        ) -> None:
             background_tasks.add_task(
                 _report_platform_usage,
                 config,
                 attempt.attempt_id,
-                "success",
-                completion.usage,
-                None,
+                outcome,
+                usage,
+                error_class,
             )
+
+        def _on_attempt_success(attempt: ResolvedAttempt) -> None:
             response.headers["X-Correlation-ID"] = attempt.attempt_id
             if rate_limit_info:
                 for key, value in rate_limit_headers(rate_limit_info).items():
                     response.headers[key] = value
-            return completion
 
-        # All attempts exhausted with retryable errors.
-        logger.error(
-            "All upstream attempts failed request_id=%s failures=%s",
-            platform_route.request_id,
-            failures,
+        return await run_platform_attempts(
+            route=platform_route,
+            attempts=attempts_to_try,
+            base_request_fields=base_request_fields,
+            run_attempt=_run_chat_attempt,
+            extract_usage=lambda completion: completion.usage,
+            classify_error=_classify_upstream_error,
+            report_attempt_outcome=_report_attempt_outcome,
+            on_success=_on_attempt_success,
+            max_tool_iterations=max_tool_iterations,
         )
-        is_single_attempt = len(attempts_to_try) <= 1
-        if last_exc is not None and isinstance(last_exc, (asyncio.TimeoutError, TimeoutError, httpx.TimeoutException)):
-            detail = "LLM provider timeout" if is_single_attempt else "All upstream providers timed out"
-            raise HTTPException(
-                status_code=status.HTTP_504_GATEWAY_TIMEOUT,
-                detail=detail,
-            ) from last_exc
-        detail = "LLM provider error" if is_single_attempt else "All upstream providers failed"
-        raise HTTPException(
-            status_code=status.HTTP_502_BAD_GATEWAY,
-            detail=detail,
-        ) from last_exc
+
+    provider, model = AnyLLM.split_model_provider(request.model)
+    provider_kwargs = get_provider_kwargs(config, provider)
 
     # Standalone path (no platform / no fallback). Same MCP / sandbox /
     # web_search semantics as platform mode above: if `mcp_servers` is set,
